Fixes various bugs in initscripts

+++ initscripts/viminit.sh
@@ -5,14 +5,18 @@ set -e
 if [ "$(command -v nvim)" ]
 then
   curl -fLo "$HOME/.config/nvim/autoload/plug.vim" --create-dirs "https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim"
-  ln -s "$HOME/dotfiles/config/nvim/init.vim" "$HOME/.config/nvim/init.vim"
+  if [ ! -e "$HOME/.config/nvim/init.vim" ]; then
+    ln -s "$HOME/dotfiles/config/nvim/init.vim" "$HOME/.config/nvim/init.vim"
+  fi
   nvim --headless +PlugInstall +qa
 fi
 
 if [ "$(command -v vim)" ]
 then
   curl -fLo "$HOME/.vim/autoload/plug.vim" --create-dirs "https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim"
-  ln -s "$HOME/dotfiles/config/nvim/init.vim" "$HOME/.vimrc"
+  if [ ! -e "$HOME/.vimrc" ]; then
+    ln -s "$HOME/dotfiles/config/nvim/init.vim" "$HOME/.vimrc"
+  fi
   vim -c 'PlugInstall | qa'
 fi
